+++ routes/apiRoutes/index.js
@@ -1,0 +1,5 @@
+const router = require('express').Router();  
+const animalRoutes = require('../apiRoutes/animalRoutes');
+
+router.use(animalRoutes);
+module.exports = router;
