+++ routes/htmlRoutes/index.js
@@ -1,0 +1,19 @@
+const path = require('path');
+const router = require('express').Router();
+
+
+
+router.get('/', (req, res) => {
+    res.sendFile(path.join(__dirname, '../../public/index.html'));
+  });
+  
+router.get('/animals', (req, res) => {
+    res.sendFile(path.join(__dirname, '../../public/animals.html'));
+  });
+  
+router.get('/zookeepers', (req, res) => {
+    res.sendFile(path.join(__dirname, '../../public/zookeepers.html'));
+  });
+  
+module.exports = router;
+
